+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   IconButton: remove the `aria-describedby` attribute when the `aria-label` and the `aria-describedby` have the same value.
+-   TextField: focus the text field input when clicking on the clear button
 
 ## [3.3.0][] - 2023-05-03